refactor: extract AgentTool to a top-level client command type
Promotes AgentTool from a type nested inside
UpdateAgentInstanceCommandStep1 to a top-level command type, so it can
be shared with the create command's builder without nesting it under
an unrelated interface. Pure move: no behavior change.
